changes to warehosue page
