+++ warehouse-inventory/src/app/warehouse1/product.ts
@@ -1,0 +1,12 @@
+export interface Product {
+  id?:string;
+  code?:string;
+  name?:string;
+  description?:string;
+  price?:number;
+  quantity?:number;
+  inventoryStatus?:string;
+  category?:string;
+  image?:string;
+  rating?:number;
+}
